test(context-budget): split #2068 regression tests under FR-006 cap; address review feedback
Address pr-review + bot-review findings on #2071:

- F1 (FR-006): context-budget.test.ts grew 1470->1960 lines (cap 500). The
  CI ratchet (check-test-file-cap.sh) silently no-ops in the quality job due
  to a shallow checkout, so CI passed but the cap was violated. Split the
  Issue #2068 T1-T15 regression block into two new files under 500 lines:
  context-budget-2068-toolpart.test.ts (T1-T7, 266) and
  context-budget-2068-edge-cases.test.ts (T8-T15, 423). Original -> 1416.
- F6: added T14 (pending-only message not masked) and T15 (short output above
  threshold still masked).
- F2: documented why the pruning estimate reads post-masking text (masking
  already credited toolMaskFreedTokens; adding it again would double-count).
- F3: release fragment test count 12 -> 15.
- Removed an orphaned estimateTokens import left by the split.
