+++ tests/unit/hooks/context-budget.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, test } from 'bun:test';
 import { createContextBudgetHandler } from '../../../src/hooks/context-budget';
-import { estimateTokens } from '../../../src/hooks/utils';
 
 // Helper function to generate messages with specific text length
 function makeMessages(textLength: number, agent?: string) {
@@ -1413,548 +1412,5 @@ describe('context-budget hook', () => {
 			expect(retrieveText).not.toContain('[Context pruned');
 			expect(retrieveText).toBe(largeText);
 		});
-	});
-});
-
-/**
- * Issue #2068 regression tests: the context-budget engine must recognize,
- * count, mask, and prune REAL OpenCode SDK tool-result payloads (`ToolPart`
- * with `part.type === 'tool'`, `part.tool`, `part.state.output`), not the
- * fictional `info.toolName` shape. These tests prove the production scenario
- * (small-context model overflow) is actually fixed.
- */
-describe('Issue #2068: real ToolPart shape', () => {
-	function baseConfig(overrides: Record<string, unknown> = {}) {
-		return {
-			context_budget: {
-				enabled: true,
-				enforce: true,
-				prune_target: 0.7,
-				recent_window: 2,
-				preserve_last_n_turns: 1,
-				tool_output_mask_threshold: 2000,
-				...overrides,
-			},
-			max_iterations: 5,
-			qa_retry_limit: 3,
-			inject_phase_reminders: true,
-		};
-	}
-
-	// T1: token counter counts state.output (usage reflects tool output).
-	test('T1: token counter counts ToolPart.state.output', async () => {
-		// 30k chars of tool output ≈ 9900 tokens. With a 10000-token limit and
-		// critical_threshold 0.9, this MUST trip critical enforcement. Before
-		// the fix, the counter ignored state.output and never tripped.
-		const config = baseConfig({ model_limits: { default: 10000 } });
-		const handler = createContextBudgetHandler(config);
-
-		const messages = [
-			toolMsg('read_file', 'x'.repeat(30000)),
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'go' }],
-			},
-		];
-		const output = { messages };
-		await handler({}, output);
-
-		// Critical enforcement should have fired and masked/pruned the tool
-		// output (proving it was counted). The architect warning is injected
-		// on the last user message.
-		const userText = output.messages[1].parts[0].text as string;
-		expect(userText).toContain('CONTEXT CRITICAL');
-	});
-
-	// T2: masking replaces ToolPart with a synthetic text placeholder.
-	test('T2: masking replaces ToolPart with text placeholder and frees tokens', async () => {
-		const config = baseConfig({
-			model_limits: { default: 10000 },
-			recent_window: 0, // force old
-			preserve_last_n_turns: 0, // allow masking the oldest message
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const big = 'x'.repeat(50000);
-		const messages = [
-			toolMsg('bash', big),
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(90000) }],
-			},
-			{
-				info: { role: 'assistant' },
-				parts: [{ type: 'text', text: 'recent assistant reply' }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		const part = output.messages[0].parts[0];
-		expect(part.type).toBe('text'); // ToolPart replaced
-		expect(part.state).toBeUndefined(); // no ToolPart anymore
-		expect(part.text as string).toMatch(/\[Tool output masked/);
-		// Freed: original placeholder is far smaller than 50000 chars.
-		expect((part.text as string).length).toBeLessThan(big.length);
-	});
-
-	// T3: recent tool output NOT masked.
-	test('T3: recent tool output NOT masked', async () => {
-		const config = baseConfig({
-			model_limits: { default: 100000 },
-			recent_window: 10, // recent
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const big = 'x'.repeat(5000);
-		const messages = [
-			toolMsg('bash', big),
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(80000) }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// age = 1 < recent_window 10 → not masked; ToolPart retained.
-		expect(output.messages[0].parts[0].type).toBe('tool');
-		expect(output.messages[0].parts[0].state.output).toBe(big);
-	});
-
-	// T4: exempt tool (read) NOT masked.
-	test('T4: exempt tool (read) NOT masked', async () => {
-		const config = baseConfig({
-			model_limits: { default: 10000 },
-			recent_window: 0,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const big = 'x'.repeat(50000);
-		const messages = [
-			toolMsg('read', big),
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(90000) }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		expect(output.messages[0].parts[0].type).toBe('tool');
-		expect(output.messages[0].parts[0].state.output).toBe(big);
-	});
-
-	// T5: full critical path on a 32k model reaches target (issue scenario).
-	test('T5: 32k-model overflow scenario frees tokens below target', async () => {
-		// Simulate the issue: a 32768-token model with ~28k tokens of tool
-		// output that previously caused "32769 tokens requested > 32768 max".
-		const config = baseConfig({
-			model_limits: { default: 32768 },
-			critical_threshold: 0.9,
-			prune_target: 0.5,
-			recent_window: 0,
-			preserve_last_n_turns: 0,
-			tool_output_mask_threshold: 100,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		// ~28k tokens ≈ 84000 chars of tool output.
-		const messages = [
-			toolMsg('bash', 'x'.repeat(84000)),
-			toolMsg('bash', 'x'.repeat(40000)),
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'proceed' }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// Recalculate post-transform tokens (text parts + completed state.output)
-		// and assert the result is at or below the configured target
-		// (32768 * 0.5 = 16384) — proving enforcement actually freed enough.
-		const target = 32768 * 0.5;
-		let postTokens = 0;
-		for (const m of output.messages) {
-			for (const p of m.parts) {
-				if (p.type === 'text' && typeof p.text === 'string') {
-					postTokens += estimateTokens(p.text);
-				} else if (
-					p.type === 'tool' &&
-					p.state?.status === 'completed' &&
-					typeof p.state.output === 'string'
-				) {
-					postTokens += estimateTokens(p.state.output);
-				}
-			}
-		}
-		expect(postTokens).toBeLessThanOrEqual(target);
-	});
-
-	// T6: idempotency — already-masked tool output not re-processed.
-	test('T6: already-masked tool output not re-masked', async () => {
-		const config = baseConfig({
-			model_limits: { default: 10000 },
-			recent_window: 0,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		// Pre-masked message (placeholder already in place as a text part).
-		const placeholder =
-			'[Tool output masked — bash returned ~100 tokens. First 200 chars: "..." Use retrieve_summary if needed.]';
-		const messages = [
-			{
-				info: { role: 'assistant' },
-				parts: [{ type: 'text', text: placeholder }],
-			},
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(90000) }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		expect(output.messages[0].parts[0].text).toBe(placeholder);
-	});
-
-	// T7: pending/running tool counted 0, not masked.
-	test('T7: pending/running tool counted 0 and not masked', async () => {
-		const config = baseConfig({
-			model_limits: { default: 100000 },
-			recent_window: 0,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const messages = [
-			{
-				info: { role: 'assistant' },
-				parts: [
-					{
-						type: 'tool',
-						tool: 'bash',
-						state: { status: 'running', input: {}, time: { start: 0 } },
-					},
-				],
-			},
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(80000) }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// Running tool has no output → not masked, shape preserved.
-		expect(output.messages[0].parts[0].type).toBe('tool');
-		expect(output.messages[0].parts[0].state.status).toBe('running');
-	});
-
-	// T8: error-state tool NOT masked (signal preserved).
-	test('T8: error-state tool output NOT masked', async () => {
-		const config = baseConfig({
-			model_limits: { default: 10000 },
-			recent_window: 0,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const errMsg = 'Error: failed to connect (stack trace ...)'.repeat(200);
-		const messages = [
-			{
-				info: { role: 'assistant' },
-				parts: [
-					{
-						type: 'tool',
-						tool: 'bash',
-						state: {
-							status: 'error',
-							input: {},
-							error: errMsg,
-							metadata: {},
-							time: { start: 0, end: 1 },
-						},
-					},
-				],
-			},
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(90000) }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// Error output preserved (not masked).
-		expect(output.messages[0].parts[0].type).toBe('tool');
-		expect(output.messages[0].parts[0].state.error).toBe(errMsg);
-	});
-
-	// T9: multiple tool parts in one message — all counted/masked independently.
-	test('T9: multiple tool parts masked independently', async () => {
-		const config = baseConfig({
-			model_limits: { default: 10000 },
-			recent_window: 0,
-			preserve_last_n_turns: 0,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const big1 = 'A'.repeat(40000);
-		const big2 = 'B'.repeat(40000);
-		const messages = [
-			{
-				info: { role: 'assistant' },
-				parts: [
-					{
-						type: 'tool',
-						tool: 'bash',
-						state: {
-							status: 'completed',
-							input: {},
-							output: big1,
-							title: 'bash',
-							metadata: {},
-							time: { start: 0, end: 1 },
-						},
-					},
-					{
-						type: 'tool',
-						tool: 'write',
-						state: {
-							status: 'completed',
-							input: {},
-							output: big2,
-							title: 'write',
-							metadata: {},
-							time: { start: 0, end: 1 },
-						},
-					},
-				],
-			},
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(90000) }],
-			},
-			{
-				info: { role: 'assistant' },
-				parts: [{ type: 'text', text: 'recent reply' }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// Both tool parts replaced with text placeholders.
-		const parts = output.messages[0].parts;
-		expect(parts.every((p) => p.type === 'text')).toBe(true);
-		expect(
-			parts.every((p) => (p.text as string).includes('[Tool output masked')),
-		).toBe(true);
-	});
-
-	// T10: tool-bearing message within preserve_last_n_turns NOT pruned or masked.
-	test('T10: recent tool message within preserve window retained intact', async () => {
-		const config = baseConfig({
-			model_limits: { default: 10000 },
-			recent_window: 0,
-			preserve_last_n_turns: 2,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const big = 'x'.repeat(50000);
-		const messages = [
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'first' }],
-			},
-			toolMsg('bash', big), // recent tool result, within preserve window
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(90000) }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// The tool message is within preserve_last_n_turns=2 → its completed
-		// ToolPart must be retained intact (not masked, not pruned).
-		const toolPart = output.messages[1].parts[0];
-		expect(toolPart.type).toBe('tool');
-		expect(toolPart.state.output).toBe(big);
-	});
-
-	// T12: mixed message — exempt + non-exempt tool parts; only non-exempt masked.
-	test('T12: mixed message masks only non-exempt tool parts', async () => {
-		// Tune so masking the one non-exempt output reaches target (no pruning,
-		// which would replace the whole message including the exempt part).
-		const config = baseConfig({
-			model_limits: { default: 12000 },
-			critical_threshold: 0.5,
-			prune_target: 0.9,
-			recent_window: 0,
-			preserve_last_n_turns: 0,
-			tool_output_mask_threshold: 100,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const exemptOutput = 'E'.repeat(12000);
-		const maskedOutput = 'M'.repeat(12000);
-		const messages = [
-			{
-				info: { role: 'assistant' },
-				parts: [
-					{
-						type: 'tool',
-						tool: 'retrieve_summary',
-						state: {
-							status: 'completed',
-							input: {},
-							output: exemptOutput,
-							title: 'retrieve_summary',
-							metadata: {},
-							time: { start: 0, end: 1 },
-						},
-					},
-					{
-						type: 'tool',
-						tool: 'bash',
-						state: {
-							status: 'completed',
-							input: {},
-							output: maskedOutput,
-							title: 'bash',
-							metadata: {},
-							time: { start: 0, end: 1 },
-						},
-					},
-				],
-			},
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'go' }],
-			},
-			{
-				info: { role: 'assistant' },
-				parts: [{ type: 'text', text: 'recent reply' }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		const parts = output.messages[0].parts;
-		// Exempt tool part retained intact.
-		const exemptPart = parts.find(
-			(p) => p.type === 'tool' && p.tool === 'retrieve_summary',
-		);
-		expect(exemptPart).toBeDefined();
-		expect(exemptPart.state.output).toBe(exemptOutput);
-		// Non-exempt tool part replaced with a masked text placeholder.
-		const maskedPart = parts.find(
-			(p) =>
-				p.type === 'text' && (p.text as string).includes('[Tool output masked'),
-		);
-		expect(maskedPart).toBeDefined();
-	});
-
-	// T11: message with sibling text part + tool part — masking replaces only the
-	// tool part; the sibling text part survives; BOTH are counted (no dedupe).
-	test('T11: text part + tool part — masking replaces only the tool part; both counted', async () => {
-		// Tune so masking the tool output alone brings usage under target, so the
-		// message is NOT subsequently pruned (which would replace the text part).
-		const config = baseConfig({
-			model_limits: { default: 8000 },
-			critical_threshold: 0.5,
-			prune_target: 0.2,
-			recent_window: 0,
-			preserve_last_n_turns: 0,
-			tool_output_mask_threshold: 100,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const big = 'x'.repeat(20000);
-		const commentary = 'assistant commentary';
-		const messages = [
-			{
-				info: { role: 'assistant' },
-				parts: [
-					{ type: 'text', text: commentary },
-					{
-						type: 'tool',
-						tool: 'bash',
-						state: {
-							status: 'completed',
-							input: {},
-							output: big,
-							title: 'bash',
-							metadata: {},
-							time: { start: 0, end: 1 },
-						},
-					},
-				],
-			},
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'go' }],
-			},
-			{
-				info: { role: 'assistant' },
-				parts: [{ type: 'text', text: 'recent reply' }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// The sibling text part is retained; the tool part is masked (replaced).
-		const parts = output.messages[0].parts;
-		const textPart = parts.find(
-			(p) => p.type === 'text' && p.text === commentary,
-		);
-		const maskedPart = parts.find(
-			(p) =>
-				p.type === 'text' && (p.text as string).includes('[Tool output masked'),
-		);
-		expect(textPart).toBeDefined();
-		expect(maskedPart).toBeDefined();
-
-		// Both parts were counted toward the budget: critical enforcement fired
-		// (usage of commentary + big tool output exceeded critical_threshold).
-		// The architect warning on the last user message proves it.
-		const userText = output.messages[1].parts[0].text as string;
-		expect(userText).toContain('CONTEXT CRITICAL');
-	});
-
-	// T13: pruning honors per-part exemptions — an exempt tool output survives
-	// even when the message is pruned (applyObservationMasking skips exempt parts).
-	test('T13: pruning preserves exempt tool output even when message is pruned', async () => {
-		// Force pruning: large user message drives usage far over target so the
-		// old tool message is removed via applyObservationMasking.
-		const config = baseConfig({
-			model_limits: { default: 4000 },
-			critical_threshold: 0.5,
-			prune_target: 0.2,
-			recent_window: 0,
-			preserve_last_n_turns: 0,
-			tool_output_mask_threshold: 100,
-		});
-		const handler = createContextBudgetHandler(config);
-
-		const exemptOutput = 'E'.repeat(3000);
-		const messages = [
-			toolMsg('retrieve_summary', exemptOutput), // exempt — must survive pruning
-			{
-				info: { role: 'user', agent: 'architect' },
-				parts: [{ type: 'text', text: 'x'.repeat(30000) }],
-			},
-			{
-				info: { role: 'assistant' },
-				parts: [{ type: 'text', text: 'reply' }],
-			},
-		];
-		const output = { messages: JSON.parse(JSON.stringify(messages)) };
-		await handler({}, output);
-
-		// The exempt tool part must retain its original output (not pruned).
-		const toolPart = output.messages[0].parts[0];
-		expect(toolPart.type).toBe('tool');
-		expect(toolPart.state.output).toBe(exemptOutput);
 	});
 });
